# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Your super  ApplicationAPI",
        "version": "1.0.0"
    },
    "paths": {
        "/pds-backend/api/getAcr": {
            "get": {
                "tags": [
                    "PDS Annual Confidential Report Management(ACR)"
                ],
                "summary": "PDS Annual Confidential Report List",
                "description": "Total Leave Type List",
                "operationId": "getAcr",
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/pds-backend/api/addacr": {
            "post": {
                "tags": [
                    "PDS Annual Confidential Report Management(ACR)"
                ],
                "summary": "Add New ACR Mgt",
                "description": "Add New ACR Mgt",
                "operationId": "addacr",
                "requestBody": {
                    "content": {
                        "multipart/form-data": {
                            "schema": {
                                "required": [],
                                "properties": {
                                    "employee_id": {
                                        "type": "text"
                                    },
                                    "emp_name": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "designation": {
                                        "type": "string",
                                        "format": "date"
                                    },
                                    "department": {
                                        "type": "text"
                                    },
                                    "office_name": {
                                        "type": "text"
                                    },
                                    "acr_year": {
                                        "type": "text"
                                    },
                                    "score": {
                                        "type": "text"
                                    },
                                    "file": {
                                        "type": "text"
                                    },
                                    "rack_number": {
                                        "type": "text"
                                    },
                                    "bill_number": {
                                        "type": "text"
                                    },
                                    "file_number": {
                                        "type": "text"
                                    },
                                    "remarks": {
                                        "type": "text"
                                    },
                                    "status": {
                                        "type": "text"
                                    }
                                },
                                "type": "object"
                            }
                        },
                        "application/json": {
                            "schema": {}
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Added ACR Mgt Setup Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/pds-backend/api/updateAcrMgt/{id}": {
            "put": {
                "tags": [
                    "PDS Annual Confidential Report Management(ACR)"
                ],
                "summary": "Update ACR Mgt Setup",
                "operationId": "updateAcrMgt",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id, eg; 1",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "employee_id": {
                                        "type": "text",
                                        "example": "2"
                                    },
                                    "emp_name": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "Yeasin"
                                    },
                                    "designation": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2"
                                    },
                                    "department": {
                                        "type": "text",
                                        "example": "2"
                                    },
                                    "office_name": {
                                        "type": "text",
                                        "example": "2"
                                    },
                                    "acr_year": {
                                        "type": "text",
                                        "example": "2023"
                                    },
                                    "score": {
                                        "type": "text",
                                        "example": "4"
                                    },
                                    "file": {
                                        "type": "text",
                                        "example": "file_update"
                                    },
                                    "rack_number": {
                                        "type": "text",
                                        "example": "2"
                                    },
                                    "bill_number": {
                                        "type": "text",
                                        "example": "2023"
                                    },
                                    "file_number": {
                                        "type": "text",
                                        "example": "4"
                                    },
                                    "remarks": {
                                        "type": "text",
                                        "example": "remark_update"
                                    },
                                    "status": {
                                        "type": "text",
                                        "example": 0
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "ACR Mgt Setup Update Successfully",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "400": {
                        "description": "Bad request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/pds-backend/api/specificAcrInfo/{id}": {
            "get": {
                "tags": [
                    "PDS Annual Confidential Report Management(ACR)"
                ],
                "summary": "Specific Acr Info Detail",
                "description": "",
                "operationId": "specificAcrInfo",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/pds-backend/api/getprofileForAcr/{id}": {
            "get": {
                "tags": [
                    "PDS Annual Confidential Report Management(ACR)"
                ],
                "summary": "Get Id Card",
                "operationId": "getprofileForAcr",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/pds-backend/api/getAcrListByEmployeeId/{id}": {
            "get": {
                "tags": [
                    "PDS Annual Confidential Report Management(ACR)"
                ],
                "summary": "Specific Acr Info Detail",
                "description": "",
                "operationId": "getAcrListByEmployeeId",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        },
                        "example": 1
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "bearer_token": {
                "type": "apiKey",
                "description": "Enter token in format (Bearer <token>)",
                "name": "Authorization",
                "in": "header"
            }
        }
    ]
}